# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Helpdesk API
  version: 0.1.0
  description: >-
    Minimal ticketing API. Uses OIDC bearer JWTs in production, or HttpOnly cookie auth in local mode.
servers:
  - url: http://localhost:8080
    description: Local API
...
security:
  - bearerAuth: []
  - cookieAuth: []
# Cross-origin requests are limited to origins set via the ALLOWED_ORIGINS env var.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [Users]
      summary: List available roles (admin)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
      security:
        - bearerAuth: []
        - cookieAuth: []
...